RFC 206: Promise: Remove cancel()/isCancelled()
The cancel method had a number of issues. It creates a race condition
with the Deferred owner. It is also in the wrong place since the Promise
is a results vessel not a unit of work. And canceling a Promise in the
middle of a Promise chain will not cancel the root Promise in the chain.
When the root Promise is resolve, an error will occur when propagating
the result up the Promise chain when the canceled Promise is reached.