promise: new PromiseFactory.resolvedWith(Promise) method
This allows the user to create a new Promise using a specific
PromiseFactory that is resolved with the specified promise. If the
user is using multiple promise factories to manage callback thread
usage, this method provides an efficient way to chain a promise
from one promise factory to a new promise from another promise
factory.

Signed-off-by: BJ Hargrave <hargrave@us.ibm.com>